requirements.txt
# Основные
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pyyaml==6.0.1

# AI / ML
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
accelerate==0.24.0  # Для ускорения на GPU
bitsandbytes==0.41.0  # Для 8-bit quantization
scikit-learn==1.3.0
pillow==10.0.1
numpy==1.24.3
pandas==2.1.0
matplotlib==3.7.2
seaborn==0.12.2  # Красивые графики

# Прокси
requests[socks]==2.31.0

# Для сохранения моделей
joblib==1.3.2